♻️ refactor: util 파일들을 util 폴더로 이동
유틸리티 함수들의 관리를 용이하게 하기 위해 util 파일들을 별도의 util 폴더로 이동시켰습니다.
